Updated to raylib 5.5.377248b

+++ lib.mod/raylib/examples/core/core_2d_camera.c
@@ -1,11 +1,13 @@
 /*******************************************************************************************
 *
-*   raylib [core] example - 2d camera
+*   raylib [core] example - 2D Camera system
 *
-*   This example has been created using raylib 1.5 (www.raylib.com)
-*   raylib is licensed under an unmodified zlib/libpng license (View raylib.h for details)
+*   Example originally created with raylib 1.5, last time updated with raylib 3.0
 *
-*   Copyright (c) 2016 Ramon Santamaria (@raysan5)
+*   Example licensed under an unmodified zlib/libpng license, which is an OSI-certified,
+*   BSD-like license that allows static linking with closed source software
+*
+*   Copyright (c) 2016-2024 Ramon Santamaria (@raysan5)
 *
 ********************************************************************************************/
 
@@ -13,6 +15,9 @@
 
 #define MAX_BUILDINGS   100
 
+//------------------------------------------------------------------------------------
+// Program main entry point
+//------------------------------------------------------------------------------------
 int main(void)
 {
     // Initialization
@@ -30,19 +35,19 @@ int main(void)
 
     for (int i = 0; i < MAX_BUILDINGS; i++)
     {
-        buildings[i].width = GetRandomValue(50, 200);
-        buildings[i].height = GetRandomValue(100, 800);
-        buildings[i].y = screenHeight - 130 - buildings[i].height;
-        buildings[i].x = -6000 + spacing;
+        buildings[i].width = (float)GetRandomValue(50, 200);
+        buildings[i].height = (float)GetRandomValue(100, 800);
+        buildings[i].y = screenHeight - 130.0f - buildings[i].height;
+        buildings[i].x = -6000.0f + spacing;
 
-        spacing += buildings[i].width;
+        spacing += (int)buildings[i].width;
 
         buildColors[i] = (Color){ GetRandomValue(200, 240), GetRandomValue(200, 240), GetRandomValue(200, 250), 255 };
     }
 
     Camera2D camera = { 0 };
-    camera.target = (Vector2){ player.x + 20, player.y + 20 };
-    camera.offset = (Vector2){ screenWidth/2, screenHeight/2 };
+    camera.target = (Vector2){ player.x + 20.0f, player.y + 20.0f };
+    camera.offset = (Vector2){ screenWidth/2.0f, screenHeight/2.0f };
     camera.rotation = 0.0f;
     camera.zoom = 1.0f;
 
@@ -54,7 +59,6 @@ int main(void)
     {
         // Update
         //----------------------------------------------------------------------------------
-        
         // Player movement
         if (IsKeyDown(KEY_RIGHT)) player.x += 2;
         else if (IsKeyDown(KEY_LEFT)) player.x -= 2;
@@ -98,8 +102,8 @@ int main(void)
 
                 DrawRectangleRec(player, RED);
 
-                DrawLine(camera.target.x, -screenHeight*10, camera.target.x, screenHeight*10, GREEN);
-                DrawLine(-screenWidth*10, camera.target.y, screenWidth*10, camera.target.y, GREEN);
+                DrawLine((int)camera.target.x, -screenHeight*10, (int)camera.target.x, screenHeight*10, GREEN);
+                DrawLine(-screenWidth*10, (int)camera.target.y, screenWidth*10, (int)camera.target.y, GREEN);
 
             EndMode2D();
